Делает функцию makeTimer и первый тест к ней
