+++ js/modules/utils/count-answer.test.js
@@ -29,41 +29,7 @@
 //   9 ответов, 2 жизни, 0 очков, игра не окончена
 
 import {assert} from 'chai';
-
-const countAnswer = (prevState, answerStatus, elapsedTime) => {
-  const newState = prevState;
-  if (answerStatus === false || elapsedTime === 0) {
-    newState.levels -= 1;
-    newState.lives -= 1;
-    if (newState.levels === 0 && newState.lives >= 0) {
-      newState.isGameEnd = true;
-      newState.isGameWin = true;
-      newState.points += prevState.lives * 50;
-    }
-    if (newState.lives < 0) {
-      newState.isGameEnd = true;
-      newState.isGameWin = false;
-    }
-  }
-
-  if (answerStatus === true) {
-    newState.points += 100;
-    newState.levels -= 1;
-    if (elapsedTime < 10) {
-      newState.points += 50;
-    }
-    if (elapsedTime > 20) {
-      newState.points -= 50;
-    }
-    if (newState.levels === 0) {
-      newState.isGameEnd = true;
-      newState.isGameWin = true;
-      newState.points += prevState.lives * 50;
-    }
-  }
-
-  return newState;
-};
+import countAnswer from './count-answer';
 
 describe(`Counting failed answer`, () => {
   it(`should correctly count wrong answer`, () => {
